feat: correções de controllers
