--- src/main/java/com/fiap/msadminapi/application/controllers/admin/produtos/store/requests/ImagemItem.java
@@ -1,4 +1,0 @@
-package com.fiap.msadminapi.application.controllers.admin.produtos.store.requests;
-
-public record ImagemItem(String nome, String url) {
-}
