+++ src/main/java/com/fiap/msadminapi/application/produtos/all/GetProdutosController.java
@@ -1,41 +1,39 @@
-package com.fiap.msadminapi.application.controllers.admin.produtos.get;
+package com.fiap.msadminapi.application.produtos.all;
 
 
 import com.fiap.msadminapi.application.response.GenericResponse;
 import com.fiap.msadminapi.application.response.PresenterResponse;
 import com.fiap.msadminapi.domain.generic.output.OutputInterface;
-import com.fiap.msadminapi.domain.output.produto.BuscaProdutoOutput;
-import com.fiap.msadminapi.domain.presenters.cliente.produto.GetProdutoPresenter;
-import com.fiap.msadminapi.domain.useCase.produto.BuscaProdutoPorUuidUseCase;
+import com.fiap.msadminapi.domain.output.produto.BuscaTodosProdutoOutput;
+import com.fiap.msadminapi.domain.presenters.cliente.produto.GetProdutosPresenter;
+import com.fiap.msadminapi.domain.useCase.produto.BuscaTodosProdutosUseCase;
 import com.fiap.msadminapi.infra.adpter.repository.produto.BuscarProdutoRepository;
 import com.fiap.msadminapi.infra.repository.ProdutoRepository;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.UUID;
-
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("admin/produto")
-public class AdminGetProdutoController {
+@RequestMapping("admin/produtos")
+public class GetProdutosController {
     private final ProdutoRepository produtoRepository;
 
-    @GetMapping("/{uuid}")
+    @GetMapping
     @Operation(tags = {"admin"})
-    public ResponseEntity<Object> getProduto(@PathVariable UUID uuid) {
-        BuscaProdutoPorUuidUseCase useCase = new BuscaProdutoPorUuidUseCase(new BuscarProdutoRepository(produtoRepository));
-        useCase.execute(uuid);
+    public ResponseEntity<Object> getAllProdutos() {
+        BuscaTodosProdutosUseCase useCase = new BuscaTodosProdutosUseCase(new BuscarProdutoRepository(produtoRepository));
+        useCase.execute();
         OutputInterface outputInterface = useCase.getBuscaProdutoOutput();
+
         if (outputInterface.getOutputStatus().getCode() != 200) {
             return new GenericResponse().response(outputInterface);
         }
 
-        GetProdutoPresenter presenter = new GetProdutoPresenter((BuscaProdutoOutput) outputInterface);
+        GetProdutosPresenter presenter = new GetProdutosPresenter((BuscaTodosProdutoOutput) outputInterface);
         return new PresenterResponse().response(presenter);
     }
 }
