+++ src/main/java/com/fiap/msadminapi/application/produtos/store/requests/ImagemItem.java
@@ -1,0 +1,4 @@
+package com.fiap.msadminapi.application.produtos.store.requests;
+
+public record ImagemItem(String nome, String url) {
+}
